#67 Rename imports.ts to deps.ts
